Fixed a number of benign memory leaks.
Turned on function tracing.

Switched nn_conv over to the MALLOC and FREE macros.

Fixed an illegal memory access in nn_conv where we were going one step over the end of the array. Note that this was hidden using the intel compiler and didn't even show up in the analysis tools but it did under MSVC!

Fixed a number of printf problems in onestage where I was giving a long int to %d and switched them over to %ld instead.

Commented out some junk in checktailmass to get rid of warnings. This really need to be fixed up properly!

Fixed an unitinialized memory read in twostage where I didn't initialize logP0 when using normalization convergence. This was also hidden in the intel compiler but stuck out like a sore thumb with MSVC.

More long int printf business in rightHandedRiemannSum.

Maybe it doesn't hang now? Heres hoping!